Group Marking!
We can now mark and unmark targets whist in a group.

I modelled the classes based on the structures i found while decompiling the client, but happy to move around if required.

The Game/Group/GroupMarkerInfo class manages all marks the group has active; each collection indexes by Marker or groupId to keep things efficient and neat (no LINQ!)

The client is smart enough to handle marks changing from target to target, so we dont need to send events, and a UnitId of 0 means "Remove this Mark".