Group Promotion.
Enables promotion Functionality.

GroupIndex is now a calculated field on the GroupMember.

I changed Members on the Group from a Dictionary to a List to maintain the order of items in the list. Dictionary<K, T> is not ordered so when we calc the index there is no guarantee we will get the same index. i.e. if a player joined, was kicked and another joined - two characters could have the same order and maintaining them could be a nightmare.