Removed Player object from the Group Member.
Removed Player object from the group member and replaced with lookups.

Some more work is needed to handle offline players as most of the time i just ignore if the leader is offline for e.g.

Reworked how the GetPlayerByCharacterId() method on the Group works, the Player object is now an output; with the return being  success result.